(chore) include a testing livekit server in docker env
one thing not mentioned a lot in the current docs is that magnify
requires a LiveKit server to talk to in order to work.

No livekit server was included so it required each developer to
understand this on their own.

Now we include a testing livekit server in the docker containers. This
makes running the app from scratch working even if we are not aware of
the internals.

the common env vars changed to match the default livekit server
settings.